+++ components/automate-cli/cmd/chef-automate/infrastructure_test.go
@@ -1,19 +1,24 @@
 package main
 
 import (
+	"context"
 	"testing"
 
+	api "github.com/chef/automate/api/interservice/deployment"
+	"github.com/chef/automate/lib/majorupgrade_utils"
+	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
+	"github.com/stretchr/testify/assert"
+	"google.golang.org/grpc"
 )
 
 var nodeDelCmd = &cobra.Command{
-	// use getClient in the command implementation
+	Use:   "node-delete [uuid]",
+	Short: "Delete node by node uuid",
+	Long:  "",
+	RunE:  runDeleteNodeCmd,
+	Args:  cobra.ExactArgs(1),
 }
-
-// var argsAws = []string{"aws"}
-// var argsExistingNodes = []string{"existing_infra"}
-// var argsSomeThingElse = []string{"something_else"}
-// var argsEmpty = []string{}
 
 func Test_runpreInfrastructureCmd(t *testing.T) {
 	tests := []struct {
@@ -22,7 +27,7 @@ func Test_runpreInfrastructureCmd(t *testing.T) {
 		args     []string
 		wantErr  bool
 	}{
-		{"Test node delete", nodeDelCmd, []string{"argsEmpty"}, true},
+		{"Test node delete", nodeDelCmd, []string{"uuid of node"}, true},
 	}
 	for _, tt := range tests {
 		t.Run(tt.testName, func(t *testing.T) {
@@ -30,5 +35,71 @@ func Test_runpreInfrastructureCmd(t *testing.T) {
 				t.Errorf("runDeleteNodeCmd() error = %v, wantErr %v", err, tt.wantErr)
 			}
 		})
+		t.Run(tt.testName, func(t *testing.T) {
+			if err := preInfrastructureCmd(tt.cmd, tt.args); (err != nil) != tt.wantErr {
+				t.Errorf("preInfrastructureCmd() error = %v, wantErr %v", err, tt.wantErr)
+			}
+		})
 	}
+}
+
+type MockDSClient struct {
+	InfrastructureNodeDeleteFunc func(ctx context.Context, in *api.InfrastructureNodeDeleteRequest, opts ...grpc.CallOption) (*api.InfrastructureNodeDeleteResponse, error)
+	CloseFunc                    func() error
+}
+
+func (mds *MockDSClient) InfrastructureNodeDelete(ctx context.Context, in *api.InfrastructureNodeDeleteRequest, opts ...grpc.CallOption) (*api.InfrastructureNodeDeleteResponse, error) {
+	return mds.InfrastructureNodeDeleteFunc(ctx, in, opts...)
+}
+
+func (mds *MockDSClient) Close() error {
+	return mds.CloseFunc()
+}
+
+func TestRunDeleteNode(t *testing.T) {
+	customWriter := majorupgrade_utils.NewCustomWriter()
+	i := &InfraFlow{
+		DsClient: &MockDSClient{InfrastructureNodeDeleteFunc: func(ctx context.Context, in *api.InfrastructureNodeDeleteRequest, opts ...grpc.CallOption) (*api.InfrastructureNodeDeleteResponse, error) {
+			return &api.InfrastructureNodeDeleteResponse{}, nil
+		}, CloseFunc: func() error {
+			return nil
+		}},
+		Writer: customWriter.CliWriter,
+	}
+	nodeId := "3d8ffe06-6281-494a-9957-34c6f3f50154"
+	err := i.RunDeleteNode(nodeId)
+	assert.Equal(t, err, nil)
+
+}
+
+func TestRunDeleteNodeFailed(t *testing.T) {
+	customWriter := majorupgrade_utils.NewCustomWriter()
+	i := &InfraFlow{
+		DsClient: &MockDSClient{InfrastructureNodeDeleteFunc: func(ctx context.Context, in *api.InfrastructureNodeDeleteRequest, opts ...grpc.CallOption) (*api.InfrastructureNodeDeleteResponse, error) {
+			return nil, errors.New("DeploymentServiceCallError")
+		}, CloseFunc: func() error {
+			return nil
+		}},
+		Writer: customWriter.CliWriter,
+	}
+	nodeId := "3d8ffe06-6281-494a-9957-34c6f3f50154"
+	err := i.RunDeleteNode(nodeId)
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "Request to delete node failed: DeploymentServiceCallError")
+}
+
+func TestRunDeleteNodeFailedForInvaliUUID(t *testing.T) {
+	customWriter := majorupgrade_utils.NewCustomWriter()
+	i := &InfraFlow{
+		DsClient: &MockDSClient{InfrastructureNodeDeleteFunc: func(ctx context.Context, in *api.InfrastructureNodeDeleteRequest, opts ...grpc.CallOption) (*api.InfrastructureNodeDeleteResponse, error) {
+			return &api.InfrastructureNodeDeleteResponse{}, nil
+		}, CloseFunc: func() error {
+			return nil
+		}},
+		Writer: customWriter.CliWriter,
+	}
+	nodeId := "not-a-valid-uuid"
+	err := i.RunDeleteNode(nodeId)
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "argument in not a valid node UUID")
 }
